requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
numpy==2.2.1
opencv-python==4.11.0.86
pillow==11.1.0
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
pyzbar==0.1.9
requests==2.32.3
urllib3==2.3.0
